+++ app/src/main/java/com/example/chatbox/startupActivity/StartupActivity.kt
@@ -9,7 +9,7 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.example.chatbox.login.LoginActivity
 import com.example.chatbox.R
-import com.example.chatbox.Signup.SignupActivity
+import com.example.chatbox.signup.SignupActivity
 import com.example.chatbox.databinding.ActivityStartupBinding
 
 class StartupActivity : AppCompatActivity() {
