+++ app/src/main/res/drawable/ic_full_divider_long.xml
@@ -1,0 +1,23 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="315dp"
+    android:height="14dp"
+    android:viewportWidth="315"
+    android:viewportHeight="14">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M0,6.5L132,6.5"
+      android:strokeAlpha="0.2"
+      android:fillColor="#00000000"
+      android:strokeColor="#CDD1D0"
+      android:fillAlpha="0.2"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M183,6.5L315,6.5"
+      android:strokeAlpha="0.2"
+      android:fillColor="#00000000"
+      android:strokeColor="#CDD1D0"
+      android:fillAlpha="0.2"/>
+  <path
+      android:pathData="M150.08,4.16C149.38,4.84 149.03,5.8 149.03,7.03C149.03,8.26 149.38,9.22 150.08,9.91C150.78,10.6 151.63,10.95 152.62,10.95C153.61,10.95 154.45,10.6 155.15,9.91C155.85,9.22 156.2,8.26 156.2,7.03C156.2,5.8 155.85,4.84 155.15,4.16C154.45,3.47 153.61,3.12 152.62,3.12C151.63,3.12 150.78,3.47 150.08,4.16ZM147.63,7.03C147.63,6.23 147.77,5.49 148.05,4.83C148.34,4.17 148.72,3.63 149.19,3.21C149.66,2.78 150.2,2.45 150.78,2.21C151.37,1.98 151.98,1.86 152.62,1.86C153.25,1.86 153.86,1.98 154.45,2.21C155.04,2.45 155.57,2.78 156.03,3.21C156.51,3.63 156.89,4.17 157.17,4.83C157.46,5.49 157.6,6.23 157.6,7.03C157.6,8.09 157.35,9.03 156.86,9.83C156.37,10.63 155.76,11.23 155.01,11.62C154.26,12.01 153.47,12.21 152.62,12.21C151.77,12.21 150.97,12.01 150.22,11.62C149.48,11.23 148.86,10.63 148.36,9.83C147.88,9.03 147.63,8.09 147.63,7.03ZM164.73,12L162.46,7.94H160.97V12H159.59V2.07H163.33C164.24,2.07 164.97,2.36 165.52,2.93C166.08,3.49 166.36,4.19 166.36,5.03C166.36,5.75 166.15,6.36 165.72,6.88C165.29,7.38 164.7,7.7 163.96,7.84L166.32,12H164.73ZM160.97,6.71H163.07C163.63,6.71 164.09,6.56 164.43,6.26C164.78,5.95 164.95,5.54 164.95,5.03C164.95,4.51 164.78,4.09 164.43,3.78C164.09,3.46 163.63,3.31 163.07,3.31H160.97V6.71Z"
+      android:fillColor="#D6E4E0"/>
+</vector>
