+++ app/src/main/res/drawable/logo_apple.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="46dp"
+    android:height="36dp"
+    android:viewportWidth="46"
+    android:viewportHeight="36">
+  <path
+      android:pathData="M32.684,24.387C32.321,25.225 31.892,25.997 31.394,26.707C30.715,27.674 30.16,28.344 29.732,28.716C29.068,29.326 28.357,29.639 27.595,29.656C27.049,29.656 26.389,29.501 25.622,29.185C24.852,28.871 24.145,28.716 23.498,28.716C22.819,28.716 22.091,28.871 21.313,29.185C20.534,29.501 19.906,29.665 19.426,29.681C18.695,29.713 17.968,29.391 17.241,28.716C16.777,28.311 16.197,27.618 15.502,26.636C14.757,25.587 14.144,24.371 13.664,22.984C13.15,21.486 12.892,20.036 12.892,18.632C12.892,17.023 13.24,15.636 13.936,14.474C14.483,13.54 15.211,12.803 16.122,12.263C17.032,11.722 18.017,11.446 19.077,11.429C19.657,11.429 20.417,11.608 21.362,11.961C22.305,12.314 22.91,12.494 23.176,12.494C23.374,12.494 24.046,12.284 25.186,11.866C26.264,11.478 27.173,11.318 27.918,11.381C29.938,11.544 31.455,12.34 32.463,13.774C30.657,14.868 29.764,16.4 29.782,18.367C29.798,19.898 30.354,21.172 31.446,22.184C31.94,22.654 32.493,23.017 33.108,23.274C32.974,23.661 32.834,24.031 32.684,24.387V24.387ZM28.053,6.164C28.053,7.364 27.615,8.485 26.741,9.522C25.686,10.755 24.41,11.468 23.027,11.355C23.008,11.204 22.999,11.052 22.999,10.901C22.999,9.748 23.5,8.515 24.391,7.507C24.836,6.996 25.401,6.572 26.087,6.233C26.772,5.9 27.419,5.715 28.028,5.684C28.046,5.844 28.053,6.005 28.053,6.164V6.164Z"
+      android:fillColor="#ffffff"/>
+</vector>
