+++ app/src/main/res/layout/activity_signup.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Signup.SignupActivity">
+    tools:context=".signup.SignupActivity">
     <ImageView
         android:id="@+id/ic_back_arrow"
         android:layout_width="24dp"
